+++ src/main/java/net/panther/endersteel/item/custom/BagelItem.java
@@ -6,8 +6,8 @@ import net.minecraft.item.Item;
 public class BagelItem extends Item {
     public BagelItem(Settings settings) {
         super(settings.food(new FoodComponent.Builder()
-                .hunger(4)         
-                .saturationModifier(0.8f) 
+                .hunger(7)
+                .saturationModifier(1.4f)
                 .build()));
     }
 }